Change design for DDS switch issue
The existing dds switch design is modified to a notification
based design. Dct now registers with SubscriptionManager for
DdsSwitch event. This change will solve DDS switch issue and
Hotswap issue.
DDS Switch Issue: Whenever a dds switch happens, the
SubscriptionManager will notify the Dct and Dct will do the
cleanupAllConnections.
Hotswap Issue: When a hotswap is performed on a subscription,
the subscription is failing to re-acquire the network. This is
resolved by notifying Dct to cleanupAllConnections whena
subscription is deactivated.

Change-Id: Idce82c387579e41c9a9ef2f9b50cb0af883c3f32
CRs-Fixed: 519221