Remove FLAG_RECEIVER_REPLACE_PENDING flag for MSIM targets
If both the SIMs broadcast ACTION_SIM_STATE_CHANGED intent with
SUB_ID, the 2nd one will replace the 1st one as the intent is
sent with FLAG_RECEIVER_REPLACE_PENDING flag. This flag is valid
in single sim scenario. However, it is not required in multi sim
cases. Hence removing this flag for Msim targets.

Change-Id: I4fe1b5744ce3feaaf6853a75598d583a3ff69e3e
CRs-Fixed: 522159