Cat: Report stkIsRunning after StkAppService is Ready
The Android framework seems to delay the intents sent to StkAppService
until the boot finishes, that is, until StkAppService receives
Boot_Complete intent from the framework. As a result before
Boot_Complete is received, StkAppService is not ready to process
the proactive commands and the proactive command times out. Also,
the CatService reports stkIsRunning to ril when the card state
moves to present without taking into account the readiness of
StkAppService.

A new api has been introduced to inform the CatService about the
readiness of the StkAppService. CatService will then report stkIsRunning
to ril only after the card state has moved to present AND
StkAppService is ready.

Change-Id: I6cb1097682062ea92bde747ab9cc5ed16104cd20
CRs-Fixed: 602191