Fix the Subscription copy from method.
While copying the subscription data, app id is not updated
if it is null for the source subscription.

This results in saving global app subscription data with
CSIM app id, when subscription is changed from CSIM to global.

Fix the copy from function in Subscription to update app Id,
app Label, app Type and icc id even when those values are null.

Change-Id: I932a2e56489c0593726624e2074197ee632dfac6
CRs-Fixed: 642188